Move subject and course related logic to seeders
