+++ app/Jobs/ProcessBatch.php
@@ -65,9 +65,6 @@ class ProcessBatch implements ShouldQueue
 
       switch($this->type){
         case 'user' : $this->payload = $self->NimbusEdu->processUser($data, $this->payload); break;
-        case 'subject' : $this->payload = $self->NimbusEdu->processSubject($data, $this->payload); break;
-        case 'coursegrade' : $this->payload = $self->NimbusEdu->processCourseGrade($data, $this->payload); break;
-        case 'curriculum' : $this->payload = $self->Syllabus->processCurriculum($data, $this->payload); break;
         case 'results' : $this->payload = $self->NimbusEdu->processResults($data, $this->payload); break;
         case 'course' : $this->payload = $self->Syllabus->processCourses($data); break;
