+++ database/seeds/DemoUsersSeeder.php
@@ -19,11 +19,6 @@ class DemoUsersSeeder extends Seeder
         'name'=> 'showers christian high school',
       ])
       ->each(function($tenant)use($count){
-          $institution = new Institution($tenant);
-          $institution->generateSubjects();
-          $institution->generateClasses();
-          $institution->generateCurriculum();
-
           factory(App\User::class, 1)->create([
             'tenant_id' => $tenant->id,
             'firstname' => 'ekama',
